backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aliyun-python-sdk-core==2.16.0",
    "fastapi==0.116.1",
    "faster-whisper==1.1.1",
    "openai==1.107.2",
    "oss2==2.19.1",
    "pydantic-settings==2.10.1",
    "python-multipart==0.0.20",
    "reportlab==4.4.3",
    "requests==2.32.3",
    "sqlalchemy==2.0.43",
    "uvicorn[standard]==0.35.0",
]
